Change to float

+++ codes/example_08/Linear.cpp
@@ -8,6 +8,6 @@ template <typename T>
 LinearSpectral2D<T>::LinearSpectral2D(basix::FiniteElement<T> element) {};
 
 NB_MODULE(Linear, m) {
-    nb::class_<LinearSpectral2D<double>>(m, "LinearSpectral2D")
-        .def(nb::init<basix::FiniteElement<double>>());
+    nb::class_<LinearSpectral2D<float>>(m, "LinearSpectral2D")
+        .def(nb::init<basix::FiniteElement<float>>());
 }
